+++ apps/core/priv/repo/migrations/20201019070828_create_projects.exs
@@ -13,7 +13,7 @@ defmodule Core.Repo.Migrations.CreateProjects do
       add :id_from_stripe_transfer, :string, null: true
       add :individual_tax_return_id, references(:individual_tax_returns, type: :uuid, on_delete: :delete_all), null: true, primary_key: false
       add :instant_matched, :boolean, null: false
-      add :offer_price, :integer, null: true
+      add :offer_price, :decimal, null: true
       add :sale_tax_id, references(:sale_taxes, type: :uuid, on_delete: :delete_all), null: true, primary_key: false
       add :service_review_id, references(:service_reviews, type: :uuid, on_delete: :delete_all), null: true, primary_key: false
       add :status, :string, null: false
